fix: Fix remaining compilation errors
- Fix int to int64 conversion in user handler
- All handlers and tests now compile successfully

+++ api/handlers/user_handler.go
@@ -124,7 +124,7 @@ func (h *UserHandler) List(c *gin.Context) {
 	// Get total count
 	total, err := h.userService.Count(c.Request.Context(), tenantID, filters)
 	if err != nil {
-		total = len(users) // Fallback
+		total = int64(len(users)) // Fallback
 	}
 
 	c.JSON(http.StatusOK, gin.H{
